Add status integration

+++ pkg/answerers/weather/weather.go
@@ -28,7 +28,7 @@ type Response struct {
 	Base string `json:"base"`
 	Main struct {
 		Temp     float64 `json:"temp"`
-		Pressure int     `json:"pressure"`
+		Pressure float64 `json:"pressure"`
 		Humidity int     `json:"humidity"`
 		TempMin  float64 `json:"temp_min"`
 		TempMax  float64 `json:"temp_max"`
@@ -36,7 +36,7 @@ type Response struct {
 	Visibility int `json:"visibility"`
 	Wind       struct {
 		Speed float64 `json:"speed"`
-		Deg   int     `json:"deg"`
+		Deg   float64 `json:"deg"`
 		Gust  float64 `json:"gust"`
 	} `json:"wind"`
 	Clouds struct {
